+++ R/aigra-imports.R
@@ -1,0 +1,4 @@
+#' @importFrom grDevices png dev.off
+#' @importFrom graphics points grid text plot.new plot.window title segments rect polygon arrows lines
+#' @importFrom utils capture.output
+NULL